Moved validate hash logic into update transaction
